+++ baseapp/baseapp.go
@@ -486,6 +486,7 @@ func (app *BaseApp) setState(mode execMode, h cmtproto.Header) {
 		ctx: sdk.NewContext(ms, h, false, app.logger).
 			WithStreamingManager(app.streamingManager).
 			WithHeaderInfo(headerInfo),
+		eventHistory: sdk.EmptyABCIEvents(),
 	}
 
 	switch mode {
@@ -782,7 +783,9 @@ func (app *BaseApp) endBlock(ctx context.Context) (sdk.EndBlock, error) {
 	var endblock sdk.EndBlock
 
 	if app.endBlocker != nil {
-		eb, err := app.endBlocker(app.finalizeBlockState.Context())
+		// Propagate the event history.
+		em := sdk.NewEventManagerWithHistory(app.finalizeBlockState.eventHistory)
+		eb, err := app.endBlocker(app.finalizeBlockState.Context().WithEventManager(em))
 		if err != nil {
 			return endblock, err
 		}
@@ -973,6 +976,7 @@ func (app *BaseApp) runTx(mode execMode, txBytes []byte) (gInfo sdk.GasInfo, res
 			}
 
 			msCache.Write()
+			app.finalizeBlockState.eventHistory = append(app.finalizeBlockState.eventHistory, result.Events...)
 		}
 
 		if len(anteEvents) > 0 && (mode == execModeFinalize || mode == execModeSimulate) {