+++ CHANGELOG.md
@@ -44,6 +44,8 @@ Ref: https://keepachangelog.com/en/1.0.0/
 * [#594](https://github.com/provenance-io/cosmos-sdk/pull/594) Provenance: Create the InputOutputCoinsProv function that's similar to InputOutputCoins but allows many inputs with a single output (as well as 1-1 and 1-many still).
 * [#595](https://github.com/provenance-io/cosmos-sdk/pull/595) Provenance: Add ability to inject a locked coins function into the bank module.
 * [#596](https://github.com/provenance-io/cosmos-sdk/pull/596) Provenance: Include the binary version in the status command output.
+* [#601](https://github.com/provenance-io/cosmos-sdk/pull/601) Provenance: Propagate the event history into the end blocker.
+* [#602](https://github.com/provenance-io/cosmos-sdk/pull/602) Provenance: Add functionality to update denom metadata via gov proposal
 * [#597](https://github.com/provenance-io/cosmos-sdk/pull/597) Provenance: Create the ReadPageRequestWithPageKeyDecoded helper.
 
 ### Improvements